+++ asyncband/src/barrier/mod.rs
@@ -56,7 +56,7 @@ use std::task::Poll;
 use crate::internal::mutex::Mutex;
 use crate::internal::waitset::WaitSet;
 use crate::internal::waitset::WakerToken;
-use crate::internal::waitset::wake_all;
+use crate::internal::wake_all;
 
 /// A synchronization primitive for multiple tasks that need to wait for each other.
 ///
