+++ asyncband/src/event/mod.rs
@@ -62,7 +62,7 @@ use std::task::Waker;
 use crate::internal::mutex::Mutex;
 use crate::internal::waitlist::WaitList;
 use crate::internal::waitlist::WaiterId;
-use crate::internal::waitset::wake_all;
+use crate::internal::wake_all;
 
 /// A reusable event that remains set until explicitly reset.
 ///
